Added optimize method and first optimizations
The :optimize() returns a new and hopefully optimized pattern.

For now, the only optimization is to merge common prefixed together
which narrow the search munch faster. It works essentially for
patterns composed of a large number of choices between character
bases sub-patterns at some point. For example:

  (P'aaa' + P'aab' + P'baa' + P'bcc'):optimize() =>
  (P'aa' * (P'a' + P'b')) + (P'b' * (P'aa' + P'cc'))

In the ideal case, the complexity passes from O(n.m) where n is input
buffer size and m pattern branches to something more like O(n.log(m)),
or even less.

Note that this will still honor pattern ordering, so to get the full
potential of this optimization, the subpatterns needs to be ordered.